requirements.txt
Flask==2.0.2
google-cloud-storage==1.43.0
Pillow
